Dev/requirements.txt
MySQL-python==1.2.3
python-memcached==1.48
Django==1.3.1
PIL==1.1.7
South==0.7.3
django-tagging==0.3.1
sorl-thumbnail==11.09.1
